tests: move wait_until() helper up to ElectrumTestCase
